added card view
Added card view and all its constituents
