+++ app/src/main/java/com/example/btonmarket/core/presentation/components/ImageThumbnail.kt
@@ -1,8 +1,10 @@
 package com.example.btonmarket.core.presentation.components
 
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
@@ -13,7 +15,8 @@ fun ImageThumbnail(resource_imageID: Int){
     Image(painter = painterResource(id = resource_imageID),
         contentDescription = "On sell item",
         modifier = Modifier
-            .size(100.dp)
-            .padding(20.dp)
+            //.size(150.dp)
+            .width(150.dp)
+            .height(200.dp)
     )
 }
