+++ app/src/main/java/com/example/btonmarket/feature_btonmarket/presentation/MainScreen.kt
@@ -1,23 +1,21 @@
 package com.example.btonmarket.feature_btonmarket.presentation
 
-import androidx.compose.foundation.layout.RowScope
 import androidx.compose.material.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.navigation.NavDestination
-import androidx.navigation.NavDestination.Companion.hierarchy
-import androidx.navigation.NavGraph.Companion.findStartDestination
-import androidx.navigation.NavHostController
-import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
 import androidx.navigation.compose.rememberNavController
-import com.example.btonmarket.feature_btonmarket.presentation.bottomNavBar.BottomBarScreen
 import com.example.btonmarket.feature_btonmarket.presentation.bottomNavBar.components.BottomNavGraph
+import  com.example.btonmarket.R
+import com.example.btonmarket.feature_btonmarket.presentation.bottomNavBar.components.BottomBar
 
 @Composable
 fun MainScreen(){
     val navController = rememberNavController()
 
     Scaffold(
+        topBar = { TopBar() },
         bottomBar = {  BottomBar(navController = navController) }
     ) {
         BottomNavGraph(navController = navController)
@@ -25,50 +23,8 @@ fun MainScreen(){
 }
 
 @Composable
-fun BottomBar(navController: NavHostController){
-    val screens = listOf(
-        BottomBarScreen.BuyScreen,
-        BottomBarScreen.CategoriesScreen,
-        BottomBarScreen.SellScreen,
-    )
-
-    val navBackStackEntry by navController.currentBackStackEntryAsState()
-    val currentDestination = navBackStackEntry?.destination
-
-    BottomNavigation {
-        screens.forEach { screen ->
-            AddItem(
-                screen=screen,
-                currentDestination=currentDestination,
-                navController=navController
-            )
-        }
+fun TopBar(){
+    TopAppBar{
+        Text(text = stringResource(id = R.string.bton_market_title), color = Color.White)
     }
-}
-
-@Composable
-fun RowScope.AddItem(
-    screen: BottomBarScreen,
-    currentDestination: NavDestination?,
-    navController: NavHostController
-){
-    BottomNavigationItem(
-        label = {
-                Text(text = screen.title)
-        },
-        icon = {
-               Icon(
-                   imageVector = screen.icon,
-                   contentDescription = "Navigation Icon"
-               )
-        },
-        selected = currentDestination?.hierarchy?.any {it.route == screen.route } == true,
-        unselectedContentColor = LocalContentColor.current.copy(alpha = ContentAlpha.disabled),
-        onClick = {
-            navController.navigate(screen.route){
-                popUpTo(navController.graph.findStartDestination().id)
-                launchSingleTop=true
-            }
-        }
-    )
 }
