+++ app/src/main/java/com/example/btonmarket/feature_btonmarket/presentation/buy/components/BuyCardView.kt
@@ -1,13 +1,18 @@
 package com.example.btonmarket.feature_btonmarket.presentation.buy.components
 
-import android.provider.Settings.Global.getString
+import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
+import androidx.compose.material.Button
 import androidx.compose.material.Card
+import androidx.compose.material.Icon
 import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ShoppingCart
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
@@ -33,11 +38,12 @@ fun BuyCardView(itemName: String,
             .width(340.dp)
             .padding(10.dp)
             .clickable { },
+        border = BorderStroke(1.dp, Color.Blue),
         elevation = 10.dp
     ) {
         Column(
             modifier = Modifier
-                .padding(15.dp)
+                .padding(15.dp, 6.dp)
                 .fillMaxWidth()
             ,
             //horizontalAlignment = Alignment.CenterHorizontally,
@@ -55,7 +61,7 @@ fun BuyCardView(itemName: String,
                 horizontalArrangement = Arrangement.SpaceAround
             ) {
                 //Text(text = "Hey")
-                ItemDetailColumn(itemName, price, negotiable)
+                ItemDetailColumn(itemName, price, condition, negotiable)
                 SellerDetailColumn(seller)
             }
 
@@ -64,33 +70,28 @@ fun BuyCardView(itemName: String,
 }
 
 @Composable
-fun ItemDetailColumn(itemName: String, price: Float, negotiable: Boolean){
-    Column() {
-        Text(
-            buildAnnotatedString {
-                //append("Item name: ")
-                withStyle(
-                    style = SpanStyle(fontWeight = FontWeight.W900, color = Color(0xFF4552B8))
-                ) {
-                    append(itemName)
-                }
-            }
-        )
+fun ItemDetailColumn(itemName: String, price: Float, condition: String, negotiable: Boolean){
+    Column{
+        Text(text = itemName, fontWeight = FontWeight.W900, color = Color(0xFF4552B8))
+        Text(text = "$condition  condition")
         Text(
             buildAnnotatedString {
 
                 withStyle(style = SpanStyle(fontWeight = FontWeight.W900)) {
                     append(price.toString())
                 }
-                append(" USD")
+                append(" USD, ")
+
+                if (negotiable) {
+                    append(stringResource(id = R.string.NEGOTIABLE))
+                    //Text(text = stringResource(id = R.string.NEGOTIABLE), fontWeight = FontWeight.Bold, color=Color.DarkGray)
+                } else {
+                    append(stringResource(id = R.string.NEGOTIABLE))
+                    //Text(text = stringResource(id = R.string.NON_NEGOTIABLE), fontWeight = FontWeight.Bold, color=Color.LightGray)
+                }
             }
         )
 
-        if (negotiable) {
-            Text(text = "Non-negotiable", fontWeight = FontWeight.Bold, color=Color.DarkGray)
-        } else {
-            Text(text = "Non-Negotiable", fontWeight = FontWeight.Bold, color=Color.LightGray)
-        }
     }
 }
 
@@ -106,6 +107,11 @@ fun SellerDetailColumn(seller: String){
                 }
             }
         )
+        Button(onClick = { /*TODO*/ }) {
+            // cart icon goes here
+            Text(text = "GET", fontWeight = FontWeight.Bold)
+            Icon(imageVector = Icons.Default.ShoppingCart, contentDescription = "Shopping cart icon", modifier = Modifier.size(12.dp))
+        }
     }
 }
