+++ app/src/main/res/values/strings.xml
@@ -2,4 +2,5 @@
     <string name="app_name">btonmarket</string>
     <string name="NEGOTIABLE">Negotiable</string>
     <string name="NON_NEGOTIABLE">Non-negotiable</string>
+    <string name="bton_market_title">Bton Market</string>
 </resources>
